This is a 2D shooting game, andIt is also the first game I made.I will update frequently and the time is uncertain
How to play? It is very simple to use the mouse to drag the joystick to control the direction of the aircraft, and Press the s, d, f, keys on the keyboard to launch missiles of three different colors. At the same time, you can also click the three buttons on the screen to launch the missiles